Finish merging ReadableStream and ReadableByteStream
