+++ reference-implementation/lib/readable-stream.js
@@ -14,8 +14,14 @@ export default class ReadableStream {
 
     this._disturbed = false;
 
+    // Initialize to undefined first because the constructor of the controller checks this
+    // variable to validate the caller.
     this._controller = undefined;
-    this._controller = new ReadableStreamController(this, underlyingSource, size, highWaterMark);
+    if (underlyingSource['pullInto'] == undefined) {
+      this._controller = new ReadableStreamController(this, underlyingSource, size, highWaterMark);
+    } else {
+      this._controller = new ReadableByteStreamController(this, underlyingSource);
+    }
   }
 
   get locked() {
@@ -36,6 +42,18 @@ export default class ReadableStream {
     }
 
     return CancelReadableStream(this, reason);
+  }
+
+  getByobReader() {
+    if (IsReadableStream(this) === false) {
+      throw new TypeError('ReadableByteStream.prototype.getByobReader can only be used on a ReadableByteStream');
+    }
+
+    if (IsReadableByteStreamController(this._controller) === false) {
+      throw new TypeError('Cannot get a ReadableStreamByobReader for a stream not constructed with a byte source');
+    }
+
+    return AcquireReadableStreamByobReader(this);
   }
 
   getReader() {
@@ -154,6 +172,190 @@ export default class ReadableStream {
   }
 }
 
+// Readers
+
+class ReadableStreamReader {
+  constructor(stream) {
+    if (IsReadableStream(stream) === false) {
+      throw new TypeError('ReadableStreamReader can only be constructed with a ReadableStream instance');
+    }
+    if (IsReadableStreamLocked(stream) === true) {
+      throw new TypeError('This stream has already been locked for exclusive reading by another reader');
+    }
+
+    InitializeReadableStreamReaderGeneric(this, stream);
+
+    this._readRequests = [];
+  }
+
+  get closed() {
+    if (IsReadableStreamReader(this) === false) {
+      return Promise.reject(
+        new TypeError('ReadableStreamReader.prototype.closed can only be used on a ReadableStreamReader'));
+    }
+
+    return this._closedPromise;
+  }
+
+  cancel(reason) {
+    if (IsReadableStreamReader(this) === false) {
+      return Promise.reject(
+        new TypeError('ReadableStreamReader.prototype.cancel can only be used on a ReadableStreamReader'));
+    }
+
+    if (this._ownerReadableStream === undefined) {
+      return Promise.reject(new TypeError('Cannot cancel a stream using a released reader'));
+    }
+
+    return CancelReadableStream(this._ownerReadableStream, reason);
+  }
+
+  read() {
+    if (IsReadableStreamReader(this) === false) {
+      return Promise.reject(
+        new TypeError('ReadableStreamReader.prototype.read can only be used on a ReadableStreamReader'));
+    }
+
+    if (this._ownerReadableStream === undefined) {
+      return Promise.reject(new TypeError('Cannot read from a released reader'));
+    }
+
+    return ReadFromReadableStreamReader(this);
+  }
+
+  releaseLock() {
+    if (IsReadableStreamReader(this) === false) {
+      throw new TypeError('ReadableStreamReader.prototype.releaseLock can only be used on a ReadableStreamReader');
+    }
+
+    if (this._ownerReadableStream === undefined) {
+      return undefined;
+    }
+
+    if (this._readRequests.length > 0) {
+      throw new TypeError('Tried to release a reader lock when that reader has pending read() calls un-settled');
+    }
+
+    if (this._ownerReadableStream._state === 'readable') {
+      this._closedPromise_reject(
+        new TypeError('Reader was released and can no longer be used to monitor the stream\'s closedness'));
+    } else {
+      this._closedPromise = Promise.reject(
+        new TypeError('Reader was released and can no longer be used to monitor the stream\'s closedness'));
+    }
+
+    ReleaseReadableStreamReaderGeneric(this);
+  }
+}
+
+class ReadableStreamByobReader {
+  constructor(stream) {
+    if (!IsReadableStream(stream)) {
+      throw new TypeError('ReadableStreamByobReader can only be constructed with a ReadableByteStream instance');
+    }
+    if (IsReadableStreamLocked(stream)) {
+      throw new TypeError('This stream has already been locked for exclusive reading by another reader');
+    }
+
+    InitializeReadableStreamReaderGeneric(this, stream);
+
+    this._readIntoRequests = [];
+  }
+
+  get closed() {
+    if (!IsReadableStreamByobReader(this)) {
+      return Promise.reject(
+        new TypeError(
+            'ReadableStreamByobReader.prototype.closed can only be used on a ReadableStreamByobReader'));
+    }
+
+    return this._closedPromise;
+  }
+
+  cancel(reason) {
+    if (!IsReadableStreamByobReader(this)) {
+      return Promise.reject(
+        new TypeError(
+            'ReadableStreamByobReader.prototype.cancel can only be used on a ReadableStreamByobReader'));
+    }
+
+    if (this._ownerReadableStream === undefined) {
+      return Promise.reject(new TypeError('Cannot cancel a stream using a released reader'));
+    }
+
+    return CancelReadableStream(this._ownerReadableStream, reason);
+  }
+
+  read(view) {
+    if (!IsReadableStreamByobReader(this)) {
+      return Promise.reject(
+        new TypeError(
+            'ReadableStreamByobReader.prototype.read can only be used on a ReadableStreamByobReader'));
+    }
+
+    if (this._ownerReadableStream === undefined) {
+      return Promise.reject(new TypeError('Cannot read from a released reader'));
+    }
+
+    const stream = this._ownerReadableStream;
+
+    stream._disturbed = true;
+
+    if (view === undefined || !ArrayBuffer.isView(view)) {
+      return Promise.reject(new TypeError('Valid view must be provided'));
+    }
+
+    const ctor = view.constructor;
+    let elementSize = 1;
+    if (ctor === Int16Array || ctor === Uint16Array || ctor === Int32Array || ctor === Uint32Array ||
+        ctor === Float32Array || ctor === Float64Array || ctor === Int8Array || ctor === Uint8Array ||
+        ctor === Uint8ClampedArray) {
+      elementSize = ctor.BYTES_PER_ELEMENT;
+    } else if (ctor !== DataView) {
+      return Promise.reject(new TypeError('view is of an unsupported type'));
+    }
+
+    if (view.byteLength === 0) {
+      return Promise.reject(new TypeError('view must have non-zero byteLength'));
+    }
+
+    if (stream._state === 'errored') {
+      return Promise.reject(stream._storedError);
+    }
+
+    // Controllers must implement this.
+    return PullFromReadableByteStreamControllerInto(
+        stream._controller, view.buffer, view.byteOffset, view.byteLength, ctor, elementSize);
+  }
+
+  releaseLock() {
+    if (!IsReadableStreamByobReader(this)) {
+      throw new TypeError(
+          'ReadableStreamByobReader.prototype.releaseLock can only be used on a ReadableStreamByobReader');
+    }
+
+    if (this._ownerReadableStream === undefined) {
+      return;
+    }
+
+    if (this._readIntoRequests.length > 0) {
+      throw new TypeError('Tried to release a reader lock when that reader has pending read() calls un-settled');
+    }
+
+    if (this._ownerReadableStream._state === 'readable') {
+      this._closedPromise_reject(
+        new TypeError('Reader was released and can no longer be used to monitor the stream\'s closedness'));
+    } else {
+      this._closedPromise = Promise.reject(
+        new TypeError('Reader was released and can no longer be used to monitor the stream\'s closedness'));
+    }
+
+    ReleaseReadableStreamReaderGeneric(this);
+  }
+}
+
+// Controllers
+
 class ReadableStreamController {
   constructor(stream, underlyingSource, size, highWaterMark) {
     if (IsReadableStream(stream) === false) {
@@ -254,305 +456,97 @@ class ReadableStreamController {
   }
 }
 
-class ReadableStreamReader {
-  constructor(stream) {
-    if (IsReadableStream(stream) === false) {
-      throw new TypeError('ReadableStreamReader can only be constructed with a ReadableStream instance');
-    }
-    if (IsReadableStreamLocked(stream) === true) {
-      throw new TypeError('This stream has already been locked for exclusive reading by another reader');
+class ReadableByteStreamController {
+  constructor(controlledReadableByteStream, underlyingByteSource) {
+    if (IsReadableStream(controlledReadableByteStream) === false) {
+      throw new TypeError('ReadableByteStreamController can only be constructed with a ReadableByteStream instance');
     }
 
-    InitializeReadableStreamReaderGeneric(this, stream);
-
-    this._readRequests = [];
-  }
-
-  get closed() {
-    if (IsReadableStreamReader(this) === false) {
-      return Promise.reject(
-        new TypeError('ReadableStreamReader.prototype.closed can only be used on a ReadableStreamReader'));
+    if (controlledReadableByteStream._controller !== undefined) {
+      throw new TypeError(
+          'ReadableByteStreamController instances can only be created by the ReadableByteStream constructor');
     }
 
-    return this._closedPromise;
+    this._cancel = CancelReadableByteStreamController;
+    this._pull = PullFromReadableByteStreamController;
+
+    this._controlledReadableByteStream = controlledReadableByteStream;
+
+    this._underlyingByteSource = underlyingByteSource;
+
+    this._pullAgain = false;
+    this._pulling = false;
+
+    this._pendingPullIntos = [];
+
+    this._queue = [];
+    this._totalQueuedBytes = 0;
+
+    this._closeRequested = false;
+
+    InvokeOrNoop(underlyingByteSource, 'start', [this]);
   }
 
-  cancel(reason) {
-    if (IsReadableStreamReader(this) === false) {
-      return Promise.reject(
-        new TypeError('ReadableStreamReader.prototype.cancel can only be used on a ReadableStreamReader'));
+  close() {
+    if (!IsReadableByteStreamController(this)) {
+      throw new TypeError(
+          'ReadableByteStreamController.prototype.close can only be used on a ReadableByteStreamController');
     }
 
-    if (this._ownerReadableStream === undefined) {
-      return Promise.reject(new TypeError('Cannot cancel a stream using a released reader'));
+    if (this._closeRequested) {
+      throw new TypeError('The stream has already been closed; do not close it again!');
+    }
+    if (this._controlledReadableByteStream._state !== 'readable') {
+      throw new TypeError('The stream is not in the readable state and cannot be closed');
     }
 
-    return CancelReadableStream(this._ownerReadableStream, reason);
+    CloseReadableByteStreamController(this);
   }
 
-  read() {
-    if (IsReadableStreamReader(this) === false) {
-      return Promise.reject(
-        new TypeError('ReadableStreamReader.prototype.read can only be used on a ReadableStreamReader'));
+  enqueue(chunk) {
+    if (!IsReadableByteStreamController(this)) {
+      throw new TypeError(
+          'ReadableByteStreamController.prototype.enqueue can only be used on a ReadableByteStreamController');
     }
 
-    if (this._ownerReadableStream === undefined) {
-      return Promise.reject(new TypeError('Cannot read from a released reader'));
+    if (this._closeRequested) {
+      throw new TypeError('stream is closed or draining');
+    }
+    if (this._controlledReadableByteStream._state !== 'readable') {
+      throw new TypeError('The stream is not in the readable state and cannot be enqueued to');
     }
 
-    return ReadFromReadableStreamReader(this);
+    EnqueueInReadableByteStreamController(this, chunk);
   }
 
-  releaseLock() {
-    if (IsReadableStreamReader(this) === false) {
-      throw new TypeError('ReadableStreamReader.prototype.releaseLock can only be used on a ReadableStreamReader');
+  error(e) {
+    if (!IsReadableByteStreamController(this)) {
+      throw new TypeError(
+          'ReadableByteStreamController.prototype.error can only be used on a ReadableByteStreamController');
     }
 
-    if (this._ownerReadableStream === undefined) {
-      return undefined;
+    if (this._controlledReadableByteStream._state !== 'readable') {
+      throw new TypeError(`The stream is ${stream._state} and so cannot be errored`);
     }
 
-    if (this._readRequests.length > 0) {
-      throw new TypeError('Tried to release a reader lock when that reader has pending read() calls un-settled');
+    ErrorReadableByteStreamController(this, e);
+  }
+
+  respond(bytesWritten, buffer) {
+    if (!IsReadableByteStreamController(this)) {
+      throw new TypeError(
+          'ReadableByteStreamController.prototype.respond can only be used on a ReadableByteStreamController');
     }
 
-    if (this._ownerReadableStream._state === 'readable') {
-      this._closedPromise_reject(
-        new TypeError('Reader was released and can no longer be used to monitor the stream\'s closedness'));
-    } else {
-      this._closedPromise = Promise.reject(
-        new TypeError('Reader was released and can no longer be used to monitor the stream\'s closedness'));
+    if (this._pendingPullIntos.length === 0) {
+      throw new TypeError('No pending BYOB read');
     }
 
-    this._ownerReadableStream._reader = undefined;
-    this._ownerReadableStream = undefined;
-
-    return undefined;
+    RespondToReadableByteStreamController(this, bytesWritten, buffer);
   }
 }
 
-
-function AcquireReadableStreamReader(stream) {
-  return new ReadableStreamReader(stream);
-}
-
-// Exposed to controllers.
-function AddReadRequestToReadableStream(stream) {
-  const readRequest = {};
-  const promise = new Promise((resolve, reject) => {
-    readRequest._resolve = resolve;
-    readRequest._reject = reject;
-  });
-
-  stream._reader._readRequests.push(readRequest);
-
-  return promise;
-}
-
-export function CancelReadableStream(stream, reason) {
-  stream._disturbed = true;
-
-  if (stream._state === 'closed') {
-    return Promise.resolve(undefined);
-  }
-  if (stream._state === 'errored') {
-    return Promise.reject(stream._storedError);
-  }
-
-  CloseReadableStream(stream);
-
-  const sourceCancelPromise = stream._controller._cancel(stream._controller, reason);
-  return sourceCancelPromise.then(() => undefined);
-}
-
-function CancelReadableStreamController(controller, reason) {
-  controller._queue = [];
-
-  return PromiseInvokeOrNoop(controller._underlyingSource, 'cancel', [reason]);
-}
-
-// A client of ReadableStreamController may use this function directly to bypass state check.
-function CloseReadableStreamController(controller) {
-  const stream = controller._controlledReadableStream;
-
-  assert(controller._closeRequested === false);
-  assert(stream._state !== 'errored');
-
-  if (stream._state === 'closed') {
-    // This will happen if the stream was closed without calling its controller's close() method, i.e. if it was closed
-    // via cancellation.
-    return undefined;
-  }
-
-  controller._closeRequested = true;
-
-  if (controller._queue.length === 0) {
-    return CloseReadableStream(stream);
-  }
-}
-
-// A client of ReadableStreamController may use this function directly to bypass state check.
-function EnqueueInReadableStreamController(controller, chunk) {
-  const stream = controller._controlledReadableStream;
-
-  assert(controller._closeRequested === false);
-  assert(stream._state !== 'errored');
-
-  if (stream._state === 'closed') {
-    // This will happen if the stream was closed without calling its controller's close() method, i.e. if it was closed
-    // via cancellation.
-    return undefined;
-  }
-
-  if (IsReadableStreamLocked(stream) === true && GetNumReadRequests(stream) > 0) {
-    RespondToReadRequest(stream, chunk);
-  } else {
-    let chunkSize = 1;
-
-    if (controller._strategySize !== undefined) {
-      try {
-        chunkSize = controller._strategySize(chunk);
-      } catch (chunkSizeE) {
-        if (stream._state === 'readable') {
-          ErrorReadableStreamController(controller, chunkSizeE);
-        }
-        throw chunkSizeE;
-      }
-    }
-
-    try {
-      EnqueueValueWithSize(controller._queue, chunk, chunkSize);
-    } catch (enqueueE) {
-      if (stream._state === 'readable') {
-        ErrorReadableStreamController(controller, enqueueE);
-      }
-      throw enqueueE;
-    }
-  }
-
-  ReadableStreamControllerCallPullIfNeeded(controller);
-
-  return undefined;
-}
-
-// Exposed to controllers.
-function ErrorReadableStream(stream, e) {
-  assert(stream._state === 'readable');
-
-  stream._storedError = e;
-  stream._state = 'errored';
-
-  const reader = stream._reader;
-
-  if (reader === undefined) {
-    return undefined;
-  }
-
-  for (const { _reject } of reader._readRequests) {
-    _reject(e);
-  }
-  reader._readRequests = [];
-
-  reader._closedPromise_reject(e);
-  reader._closedPromise_resolve = undefined;
-  reader._closedPromise_reject = undefined;
-}
-
-// A client of ReadableStreamController may use this function directly to bypass state check.
-function ErrorReadableStreamController(controller, e) {
-  controller._queue = [];
-
-  ErrorReadableStream(controller._controlledReadableStream, e);
-}
-
-// Exposed to controllers.
-function GetNumReadRequests(stream) {
-  return stream._reader._readRequests.length;
-}
-
-// Exposed to controllers.
-export function CloseReadableStream(stream) {
-  assert(stream._state === 'readable');
-
-  stream._state = 'closed';
-
-  const reader = stream._reader;
-
-  if (reader === undefined) {
-    return undefined;
-  }
-
-  if (IsReadableStreamReader(reader)) {
-    for (const { _resolve } of reader._readRequests) {
-      _resolve(CreateIterResultObject(undefined, true));
-    }
-    reader._readRequests = [];
-  }
-
-  CloseReadableStreamReaderGeneric(reader);
-
-  return undefined;
-}
-
-export function CloseReadableStreamReaderGeneric(reader) {
-  reader._closedPromise_resolve(undefined);
-  reader._closedPromise_resolve = undefined;
-  reader._closedPromise_reject = undefined;
-}
-
-// A client of ReadableStreamController may use this function directly to bypass state check.
-function GetReadableStreamControllerDesiredSize(controller) {
-  const queueSize = GetTotalQueueSize(controller._queue);
-  return controller._strategyHWM - queueSize;
-}
-
-export function InitializeReadableStreamReaderGeneric(reader, stream) {
-  reader._ownerReadableStream = stream;
-  stream._reader = reader;
-
-  if (stream._state === 'readable') {
-    reader._closedPromise = new Promise((resolve, reject) => {
-      reader._closedPromise_resolve = resolve;
-      reader._closedPromise_reject = reject;
-    });
-  } else {
-    if (stream._state === 'closed') {
-      reader._closedPromise = Promise.resolve(undefined);
-      reader._closedPromise_resolve = undefined;
-      reader._closedPromise_reject = undefined;
-    } else {
-      assert(stream._state === 'errored', 'state must be errored');
-
-      reader._closedPromise = Promise.reject(stream._storedError);
-      reader._closedPromise_resolve = undefined;
-      reader._closedPromise_reject = undefined;
-    }
-  }
-}
-
-export function IsReadableByteStreamReader(x) {
-  if (!typeIsObject(x)) {
-    return false;
-  }
-
-  if (!Object.prototype.hasOwnProperty.call(x, '_readRequests')) {
-    return false;
-  }
-
-  return true;
-}
-
-export function IsReadableStream(x) {
-  if (!typeIsObject(x)) {
-    return false;
-  }
-
-  if (!Object.prototype.hasOwnProperty.call(x, '_controller')) {
-    return false;
-  }
-
-  return true;
-}
+// Utilities
 
 export function IsReadableStreamDisturbed(stream) {
   assert(IsReadableStream(stream) === true, 'IsReadableStreamDisturbed should only be used on known readable streams');
@@ -560,153 +554,14 @@ export function IsReadableStreamDisturbed(stream) {
   return stream._disturbed;
 }
 
-// Exposed to controllers.
-function IsReadableStreamLocked(stream) {
-  assert(IsReadableStream(stream) === true, 'IsReadableStreamLocked should only be used on known readable streams');
+// Abstract operations for the ReadableStream.
 
-  if (stream._reader === undefined) {
-    return false;
-  }
-
-  return true;
+function AcquireReadableStreamByobReader(stream) {
+  return new ReadableStreamByobReader(stream);
 }
 
-function IsReadableStreamController(x) {
-  if (!typeIsObject(x)) {
-    return false;
-  }
-
-  if (!Object.prototype.hasOwnProperty.call(x, '_controlledReadableStream')) {
-    return false;
-  }
-
-  return true;
-}
-
-function IsReadableStreamReader(x) {
-  if (!typeIsObject(x)) {
-    return false;
-  }
-
-  if (!Object.prototype.hasOwnProperty.call(x, '_readRequests')) {
-    return false;
-  }
-
-  return true;
-}
-
-function PullFromReadableStreamController(controller) {
-  const stream = controller._controlledReadableStream;
-
-  if (controller._queue.length > 0) {
-    const chunk = DequeueValue(controller._queue);
-
-    if (controller._closeRequested === true && controller._queue.length === 0) {
-      CloseReadableStream(stream);
-    } else {
-      ReadableStreamControllerCallPullIfNeeded(controller);
-    }
-
-    return Promise.resolve(CreateIterResultObject(chunk, false));
-  }
-
-  const pendingPromise = AddReadRequestToReadableStream(stream);
-  ReadableStreamControllerCallPullIfNeeded(controller);
-  return pendingPromise;
-}
-
-function ReadFromReadableStreamReader(reader) {
-  const stream = reader._ownerReadableStream;
-
-  assert(stream !== undefined);
-
-  stream._disturbed = true;
-
-  if (stream._state === 'closed') {
-    return Promise.resolve(CreateIterResultObject(undefined, true));
-  }
-
-  if (stream._state === 'errored') {
-    return Promise.reject(stream._storedError);
-  }
-
-  assert(stream._state === 'readable');
-
-  // Controllers must implement this.
-  return stream._controller._pull(stream._controller);
-}
-
-function ReadableStreamControllerCallPullIfNeeded(controller) {
-  const shouldPull = ShouldPullReadableStreamController(controller);
-  if (shouldPull === false) {
-    return undefined;
-  }
-
-  if (controller._pulling === true) {
-    controller._pullAgain = true;
-    return undefined;
-  }
-
-  controller._pulling = true;
-
-  const stream = controller._controlledReadableStream;
-
-  const pullPromise = PromiseInvokeOrNoop(controller._underlyingSource, 'pull', [controller]);
-  pullPromise.then(
-    () => {
-      controller._pulling = false;
-
-      if (controller._pullAgain === true) {
-        controller._pullAgain = false;
-        return ReadableStreamControllerCallPullIfNeeded(controller);
-      }
-    },
-    e => {
-      if (stream._state === 'readable') {
-        return ErrorReadableStreamController(controller, e);
-      }
-    }
-  )
-  .catch(rethrowAssertionErrorRejection);
-
-  return undefined;
-}
-
-// Exposed to controllers.
-function RespondToReadRequest(stream, chunk) {
-  const reader = stream._reader;
-
-  assert(reader._readRequests.length > 0);
-
-  const readRequest = reader._readRequests.shift();
-  readRequest._resolve(CreateIterResultObject(chunk, false));
-}
-
-function ShouldPullReadableStreamController(controller) {
-  const stream = controller._controlledReadableStream;
-
-  if (stream._state === 'closed' || stream._state === 'errored') {
-    return false;
-  }
-
-  if (controller._closeRequested === true) {
-    return false;
-  }
-
-  if (controller._started === false) {
-    return false;
-  }
-
-  if (IsReadableStreamLocked(stream) === true && GetNumReadRequests(stream) > 0) {
-    return true;
-  }
-
-  const desiredSize = GetReadableStreamControllerDesiredSize(controller);
-  if (desiredSize > 0) {
-    return true;
-  }
-
-  return false;
+function AcquireReadableStreamReader(stream) {
+  return new ReadableStreamReader(stream);
 }
 
 function TeeReadableStream(stream, shouldClone) {
@@ -837,4 +692,871 @@ function create_TeeReadableStreamBranch2CancelFunction() {
     return teeState.promise;
   };
   return f;
+}
+
+// Type checking helpers.
+
+function IsReadableStream(x) {
+  if (!typeIsObject(x)) {
+    return false;
+  }
+
+  if (!Object.prototype.hasOwnProperty.call(x, '_controller')) {
+    return false;
+  }
+
+  return true;
+}
+
+function IsReadableStreamReader(x) {
+  if (!typeIsObject(x)) {
+    return false;
+  }
+
+  if (!Object.prototype.hasOwnProperty.call(x, '_readRequests')) {
+    return false;
+  }
+
+  return true;
+}
+
+function IsReadableStreamByobReader(x) {
+  if (!typeIsObject(x)) {
+    return false;
+  }
+
+  if (!Object.prototype.hasOwnProperty.call(x, '_readIntoRequests')) {
+    return false;
+  }
+
+  return true;
+}
+
+function IsReadableStreamController(x) {
+  if (!typeIsObject(x)) {
+    return false;
+  }
+
+  if (!Object.prototype.hasOwnProperty.call(x, '_controlledReadableStream')) {
+    return false;
+  }
+
+  return true;
+}
+
+function IsReadableByteStreamController(x) {
+  if (!typeIsObject(x)) {
+    return false;
+  }
+
+  if (!Object.prototype.hasOwnProperty.call(x, '_controlledReadableByteStream')) {
+    return false;
+  }
+
+  return true;
+}
+
+// ReadableStream API exposed for controllers.
+
+function AddReadRequestToReadableStream(stream) {
+  const readRequest = {};
+  const promise = new Promise((resolve, reject) => {
+    readRequest._resolve = resolve;
+    readRequest._reject = reject;
+  });
+
+  stream._reader._readRequests.push(readRequest);
+
+  return promise;
+}
+
+function AddReadIntoRequestToReadableStream(stream, byteOffset, byteLength, ctor, elementSize) {
+  return new Promise((resolve, reject) => {
+    const req = {
+      resolve,
+      reject,
+      byteOffset,
+      byteLength,
+      ctor,
+      elementSize
+    };
+    stream._reader._readIntoRequests.push(req);
+  });
+}
+
+function CloseReadableStream(stream) {
+  assert(stream._state === 'readable');
+
+  stream._state = 'closed';
+
+  const reader = stream._reader;
+
+  if (reader === undefined) {
+    return undefined;
+  }
+
+  if (IsReadableStreamReader(reader)) {
+    for (const { _resolve } of reader._readRequests) {
+      _resolve(CreateIterResultObject(undefined, true));
+    }
+    reader._readRequests = [];
+  }
+
+  reader._closedPromise_resolve(undefined);
+  reader._closedPromise_resolve = undefined;
+  reader._closedPromise_reject = undefined;
+
+  return undefined;
+}
+
+function ErrorReadableStream(stream, e) {
+  assert(IsReadableStream(stream), 'stream must be ReadableByteStream');
+  assert(stream._state === 'readable', 'state must be readable');
+
+  stream._state = 'errored';
+  stream._storedError = e;
+
+  const reader = stream._reader;
+
+  if (reader === undefined) {
+    return undefined;
+  }
+
+  if (IsReadableStreamReader(reader)) {
+    for (const req of reader._readRequests) {
+      req._reject(e);
+    }
+
+    reader._readRequests = [];
+  } else {
+    assert(IsReadableStreamByobReader(reader), 'reader must be ReadableStreamByobReader');
+
+    for (const req of reader._readIntoRequests) {
+      req.reject(e);
+    }
+
+    reader._readIntoRequests = [];
+  }
+
+  reader._closedPromise_reject(e);
+  reader._closedPromise_resolve = undefined;
+  reader._closedPromise_reject = undefined;
+}
+
+function GetNumReadRequests(stream) {
+  return stream._reader._readRequests.length;
+}
+
+function GetNumReadIntoRequests(stream) {
+  return stream._reader._readIntoRequests.length;
+}
+
+function IsReadableStreamLocked(stream) {
+  assert(IsReadableStream(stream) === true, 'IsReadableStreamLocked should only be used on known readable streams');
+
+  if (stream._reader === undefined) {
+    return false;
+  }
+
+  return true;
+}
+
+function ReadableStreamHasByobReader(stream) {
+  const reader = stream._reader;
+
+  return reader !== undefined && IsReadableStreamByobReader(reader);
+}
+
+function ReadableStreamHasReader(stream) {
+  const reader = stream._reader;
+
+  return reader != undefined && IsReadableStreamReader(reader);
+}
+
+function ReleaseReadableStreamReaderGenericForController(stream) {
+  ReleaseReadableStreamReaderGeneric(stream._reader);
+}
+
+function RespondToReadIntoRequest(stream, buffer, length) {
+  const reader = stream._reader;
+
+  assert(reader._readIntoRequests.length > 0,
+         'readIntoRequest must not be empty when calling RespondToReadIntoRequest');
+  assert(stream._state !== 'errored', 'state must not be errored');
+
+  const req = reader._readIntoRequests.shift();
+  const ctor = req.ctor;
+  const byteOffset = req.byteOffset;
+
+  if (stream._state === 'closed') {
+    assert(length === undefined);
+    const view = new ctor(buffer, byteOffset, 0);
+    req.resolve(CreateIterResultObject(view, true));
+
+    return;
+  }
+
+  assert(length <= req.byteLength);
+  assert(length % req.elementSize === 0);
+
+  const view = new ctor(buffer, byteOffset, length / req.elementSize);
+  req.resolve(CreateIterResultObject(view, false));
+}
+
+function RespondToReadRequest(stream, chunk) {
+  const reader = stream._reader;
+
+  assert(reader._readRequests.length > 0);
+
+  const readRequest = reader._readRequests.shift();
+  readRequest._resolve(CreateIterResultObject(chunk, false));
+}
+
+// Abstract operations for the readers.
+
+// A client of ReadableStreamReader may use this function directly to bypass state check.
+function CancelReadableStream(stream, reason) {
+  stream._disturbed = true;
+
+  if (stream._state === 'closed') {
+    return Promise.resolve(undefined);
+  }
+  if (stream._state === 'errored') {
+    return Promise.reject(stream._storedError);
+  }
+
+  CloseReadableStream(stream);
+
+  const sourceCancelPromise = stream._controller._cancel(stream._controller, reason);
+  return sourceCancelPromise.then(() => undefined);
+}
+
+// A client of ReadableStreamReader may use this function directly to bypass state check.
+function ReadFromReadableStreamReader(reader) {
+  const stream = reader._ownerReadableStream;
+
+  assert(stream !== undefined);
+
+  stream._disturbed = true;
+
+  if (stream._state === 'closed') {
+    return Promise.resolve(CreateIterResultObject(undefined, true));
+  }
+
+  if (stream._state === 'errored') {
+    return Promise.reject(stream._storedError);
+  }
+
+  assert(stream._state === 'readable');
+
+  return stream._controller._pull(stream._controller);
+}
+
+function InitializeReadableStreamReaderGeneric(reader, stream) {
+  reader._ownerReadableStream = stream;
+  stream._reader = reader;
+
+  if (stream._state === 'readable') {
+    reader._closedPromise = new Promise((resolve, reject) => {
+      reader._closedPromise_resolve = resolve;
+      reader._closedPromise_reject = reject;
+    });
+  } else {
+    if (stream._state === 'closed') {
+      reader._closedPromise = Promise.resolve(undefined);
+      reader._closedPromise_resolve = undefined;
+      reader._closedPromise_reject = undefined;
+    } else {
+      assert(stream._state === 'errored', 'state must be errored');
+
+      reader._closedPromise = Promise.reject(stream._storedError);
+      reader._closedPromise_resolve = undefined;
+      reader._closedPromise_reject = undefined;
+    }
+  }
+}
+
+function ReleaseReadableStreamReaderGeneric(reader) {
+  assert(reader._ownerReadableStream._reader !== undefined);
+  assert(reader._ownerReadableStream !== undefined);
+
+  reader._ownerReadableStream._reader = undefined;
+  reader._ownerReadableStream = undefined;
+}
+
+// Abstract operations for the ReadableStreamController.
+
+// A client of ReadableStreamController may use this function directly to bypass state check.
+function CloseReadableStreamController(controller) {
+  const stream = controller._controlledReadableStream;
+
+  assert(controller._closeRequested === false);
+  assert(stream._state !== 'errored');
+
+  if (stream._state === 'closed') {
+    // This will happen if the stream was closed without calling its controller's close() method, i.e. if it was closed
+    // via cancellation.
+    return undefined;
+  }
+
+  controller._closeRequested = true;
+
+  if (controller._queue.length === 0) {
+    return CloseReadableStream(stream);
+  }
+}
+
+// A client of ReadableStreamController may use this function directly to bypass state check.
+function GetReadableStreamControllerDesiredSize(controller) {
+  const queueSize = GetTotalQueueSize(controller._queue);
+  return controller._strategyHWM - queueSize;
+}
+
+// A client of ReadableStreamController may use this function directly to bypass state check.
+function EnqueueInReadableStreamController(controller, chunk) {
+  const stream = controller._controlledReadableStream;
+
+  assert(controller._closeRequested === false);
+  assert(stream._state !== 'errored');
+
+  if (stream._state === 'closed') {
+    // This will happen if the stream was closed without calling its controller's close() method, i.e. if it was closed
+    // via cancellation.
+    return undefined;
+  }
+
+  if (IsReadableStreamLocked(stream) === true && GetNumReadRequests(stream) > 0) {
+    RespondToReadRequest(stream, chunk);
+  } else {
+    let chunkSize = 1;
+
+    if (controller._strategySize !== undefined) {
+      try {
+        chunkSize = controller._strategySize(chunk);
+      } catch (chunkSizeE) {
+        if (stream._state === 'readable') {
+          ErrorReadableStreamController(controller, chunkSizeE);
+        }
+        throw chunkSizeE;
+      }
+    }
+
+    try {
+      EnqueueValueWithSize(controller._queue, chunk, chunkSize);
+    } catch (enqueueE) {
+      if (stream._state === 'readable') {
+        ErrorReadableStreamController(controller, enqueueE);
+      }
+      throw enqueueE;
+    }
+  }
+
+  ReadableStreamControllerCallPullIfNeeded(controller);
+
+  return undefined;
+}
+
+// A client of ReadableStreamController may use this function directly to bypass state check.
+function ErrorReadableStreamController(controller, e) {
+  controller._queue = [];
+
+  ErrorReadableStream(controller._controlledReadableStream, e);
+}
+
+function CancelReadableStreamController(controller, reason) {
+  controller._queue = [];
+
+  return PromiseInvokeOrNoop(controller._underlyingSource, 'cancel', [reason]);
+}
+
+function PullFromReadableStreamController(controller) {
+  const stream = controller._controlledReadableStream;
+
+  if (controller._queue.length > 0) {
+    const chunk = DequeueValue(controller._queue);
+
+    if (controller._closeRequested === true && controller._queue.length === 0) {
+      CloseReadableStream(stream);
+    } else {
+      ReadableStreamControllerCallPullIfNeeded(controller);
+    }
+
+    return Promise.resolve(CreateIterResultObject(chunk, false));
+  }
+
+  const pendingPromise = AddReadRequestToReadableStream(stream);
+  ReadableStreamControllerCallPullIfNeeded(controller);
+  return pendingPromise;
+}
+
+function ReadableStreamControllerCallPullIfNeeded(controller) {
+  const shouldPull = ShouldPullReadableStreamController(controller);
+  if (shouldPull === false) {
+    return undefined;
+  }
+
+  if (controller._pulling === true) {
+    controller._pullAgain = true;
+    return undefined;
+  }
+
+  controller._pulling = true;
+
+  const stream = controller._controlledReadableStream;
+
+  const pullPromise = PromiseInvokeOrNoop(controller._underlyingSource, 'pull', [controller]);
+  pullPromise.then(
+    () => {
+      controller._pulling = false;
+
+      if (controller._pullAgain === true) {
+        controller._pullAgain = false;
+        return ReadableStreamControllerCallPullIfNeeded(controller);
+      }
+    },
+    e => {
+      if (stream._state === 'readable') {
+        return ErrorReadableStreamController(controller, e);
+      }
+    }
+  )
+  .catch(rethrowAssertionErrorRejection);
+
+  return undefined;
+}
+
+function ShouldPullReadableStreamController(controller) {
+  const stream = controller._controlledReadableStream;
+
+  if (stream._state === 'closed' || stream._state === 'errored') {
+    return false;
+  }
+
+  if (controller._closeRequested === true) {
+    return false;
+  }
+
+  if (controller._started === false) {
+    return false;
+  }
+
+  if (IsReadableStreamLocked(stream) === true && GetNumReadRequests(stream) > 0) {
+    return true;
+  }
+
+  const desiredSize = GetReadableStreamControllerDesiredSize(controller);
+  if (desiredSize > 0) {
+    return true;
+  }
+
+  return false;
+}
+
+// Abstract operations for the ReadableByteStreamController.
+
+// A client of ReadableByteStreamController may use this function directly to bypass state check.
+function CloseReadableByteStreamController(controller) {
+  const stream = controller._controlledReadableByteStream;
+
+  if (controller._totalQueuedBytes > 0) {
+    controller._closeRequested = true;
+
+    return;
+  }
+
+  if (ReadableStreamHasByobReader(stream) &&
+      controller._pendingPullIntos.length > 0 &&
+      controller._pendingPullIntos[0].bytesFilled > 0) {
+    const e = new TypeError('Insufficient bytes to fill elements in the given buffer');
+    ErrorReadableByteStreamController(controller, e);
+
+    throw e;
+  }
+
+  CloseReadableStream(stream);
+}
+
+// A client of ReadableByteStreamController may use this function directly to bypass state check.
+function EnqueueInReadableByteStreamController(controller, chunk) {
+  const stream = controller._controlledReadableByteStream;
+
+  const buffer = chunk.buffer;
+  const byteOffset = chunk.byteOffset;
+  const byteLength = chunk.byteLength;
+
+  if (ReadableStreamHasReader(stream)) {
+    if (GetNumReadRequests(stream) === 0) {
+      EnqueueChunkToQueueOfController(controller, TransferArrayBuffer(buffer), byteOffset, byteLength);
+    } else {
+      assert(controller._queue.length === 0);
+
+      const transferredView = new Uint8Array(TransferArrayBuffer(buffer), byteOffset, byteLength);
+      RespondToReadRequest(stream, transferredView);
+
+      if (GetNumReadRequests(stream) > 0) {
+        ReadableByteStreamControllerCallPullOrPullIntoLaterIfNeeded(controller);
+      }
+    }
+  } else if (ReadableStreamHasByobReader(stream)) {
+    // TODO: Ideally this detaching should happen only if the buffer is not consumed fully.
+    EnqueueChunkToQueueOfController(controller, TransferArrayBuffer(buffer), byteOffset, byteLength);
+    RespondToReadIntoRequestsFromQueue(controller);
+  } else {
+    assert(!IsReadableStreamLocked(stream), 'stream must not be locked');
+    EnqueueChunkToQueueOfController(controller, TransferArrayBuffer(buffer), byteOffset, byteLength);
+  }
+}
+
+// A client of ReadableByteStreamController may use this function directly to bypass state check.
+function ErrorReadableByteStreamController(controller, e) {
+  controller._pendingPullIntos = []
+  controller._queue = [];
+  ErrorReadableStream(controller._controlledReadableByteStream, e);
+}
+
+// A client of ReadableByteStreamController may use this function directly to bypass state check.
+function RespondToReadableByteStreamController(controller, bytesWritten, buffer) {
+  const stream = controller._controlledReadableByteStream;
+
+  assert(ReadableStreamHasByobReader(stream), 'reader must be ReadableStreamByobReader');
+
+  if (stream._state === 'closed') {
+    if (bytesWritten !== 0) {
+      throw new TypeError('bytesWritten must be 0 when calling respond() on a closed stream');
+    }
+
+    RespondToByobReaderInClosedState(controller, buffer);
+  } else {
+    assert(stream._state === 'readable');
+
+    RespondToByobReaderInReadableState(controller, bytesWritten, buffer);
+  }
+}
+
+function CancelReadableByteStreamController(controller, reason) {
+  if (controller._pendingPullIntos.length > 0) {
+    controller._pendingPullIntos[0].bytesFilled = 0;
+  }
+
+  controller._queue = [];
+  controller._totalQueuedBytes = 0;
+
+  return PromiseInvokeOrNoop(controller._underlyingByteSource, 'cancel', [reason]);
+}
+
+function EnqueueChunkToQueueOfController(controller, buffer, byteOffset, byteLength) {
+  controller._queue.push({buffer, byteOffset, byteLength});
+  controller._totalQueuedBytes += byteLength;
+}
+
+function FillPullIntoDescriptorFromQueue(controller, pullIntoDescriptor) {
+  const elementSize = pullIntoDescriptor.elementSize;
+
+  const currentAlignedBytes = pullIntoDescriptor.bytesFilled - pullIntoDescriptor.bytesFilled % elementSize;
+
+  const maxBytesToCopy = Math.min(controller._totalQueuedBytes,
+                                  pullIntoDescriptor.byteLength - pullIntoDescriptor.bytesFilled);
+  const maxBytesFilled = pullIntoDescriptor.bytesFilled + maxBytesToCopy;
+  const maxAlignedBytes = maxBytesFilled - maxBytesFilled % elementSize;
+
+  let totalBytesToCopyRemaining = maxBytesToCopy;
+  let ready = false;
+  if (maxAlignedBytes > currentAlignedBytes) {
+    totalBytesToCopyRemaining = maxAlignedBytes - pullIntoDescriptor.bytesFilled;
+    ready = true;
+  }
+
+  const queue = controller._queue;
+
+  while (totalBytesToCopyRemaining > 0) {
+    const headOfQueue = queue[0];
+
+    const bytesToCopy = Math.min(totalBytesToCopyRemaining, headOfQueue.byteLength);
+
+    const destStart = pullIntoDescriptor.byteOffset + pullIntoDescriptor.bytesFilled;
+    new Uint8Array(pullIntoDescriptor.buffer).set(
+        new Uint8Array(headOfQueue.buffer, headOfQueue.byteOffset, bytesToCopy), destStart);
+
+    if (headOfQueue.byteLength === bytesToCopy) {
+      queue.shift();
+    } else {
+      headOfQueue.byteOffset += bytesToCopy;
+      headOfQueue.byteLength -= bytesToCopy;
+    }
+    controller._totalQueuedBytes -= bytesToCopy;
+
+    pullIntoDescriptor.bytesFilled += bytesToCopy;
+
+    totalBytesToCopyRemaining -= bytesToCopy
+  }
+
+  if (!ready) {
+    assert(controller._totalQueuedBytes === 0, 'queue must be empty');
+    assert(pullIntoDescriptor.bytesFilled > 0);
+    assert(pullIntoDescriptor.bytesFilled < pullIntoDescriptor.elementSize);
+  }
+
+  return ready;
+}
+
+function PullFromReadableByteStreamController(controller) {
+  const stream = controller._controlledReadableByteStream;
+
+  if (GetNumReadRequests(stream) >= 1) {
+    const pendingPromise = AddReadRequestToReadableStream(stream);
+    return pendingPromise;
+  }
+
+  if (controller._totalQueuedBytes > 0) {
+    const entry = controller._queue.shift();
+    controller._totalQueuedBytes -= entry.byteLength;
+
+    const view = new Uint8Array(entry.buffer, entry.byteOffset, entry.byteLength);
+    const promise = Promise.resolve(CreateIterResultObject(view, false));
+
+    if (controller._totalQueuedBytes === 0 && controller._closeRequested) {
+      CloseReadableStream(stream);
+    }
+
+    return promise;
+  }
+
+  const promise = AddReadRequestToReadableStream(stream);
+
+  if (controller._pulling) {
+    controller._pullAgain = true;
+    return promise;
+  }
+
+  ReadableByteStreamControllerCallPull(controller);
+  ReadableByteStreamControllerCallPullOrPullIntoRepeatedlyIfNeeded(controller);
+
+  return promise;
+}
+
+function PullFromReadableByteStreamControllerInto(controller, buffer, byteOffset, byteLength, ctor, elementSize) {
+  const stream = controller._controlledReadableByteStream;
+
+  const pullIntoDescriptor = {
+    buffer,
+    byteOffset,
+    byteLength,
+    bytesFilled: 0,
+    elementSize
+  };
+
+  if (controller._pendingPullIntos.length > 0) {
+    pullIntoDescriptor.buffer = TransferArrayBuffer(pullIntoDescriptor.buffer);
+    controller._pendingPullIntos.push(pullIntoDescriptor);
+
+    return AddReadIntoRequestToReadableStream(stream, byteOffset, byteLength, ctor, elementSize);
+  }
+
+  if (stream._state === 'closed') {
+    return Promise.resolve(CreateIterResultObject(new ctor(buffer, byteOffset, 0), true));
+  }
+
+  if (controller._totalQueuedBytes > 0) {
+    const ready = FillPullIntoDescriptorFromQueue(controller, pullIntoDescriptor);
+
+    if (ready) {
+      assert(pullIntoDescriptor.bytesFilled <= byteLength);
+      assert(pullIntoDescriptor.bytesFilled % elementSize === 0);
+
+      const view = new ctor(buffer, byteOffset, pullIntoDescriptor.bytesFilled / elementSize);
+      const promise = Promise.resolve(CreateIterResultObject(view, false));
+
+      if (controller._totalQueuedBytes === 0 && controller._closeRequested) {
+        CloseReadableStream(stream);
+      }
+
+      return promise;
+    }
+
+    if (controller._closeRequested) {
+      const e = new TypeError('Insufficient bytes to fill elements in the given buffer');
+      ErrorReadableByteStreamController(controller, e)
+
+      return Promise.reject(e);
+    }
+  }
+
+  pullIntoDescriptor.buffer = TransferArrayBuffer(pullIntoDescriptor.buffer);
+  controller._pendingPullIntos.push(pullIntoDescriptor);
+
+  const promise = AddReadIntoRequestToReadableStream(stream, byteOffset, byteLength, ctor, elementSize);
+
+  if (controller._pulling) {
+    controller._pullAgain = true;
+    return promise;
+  }
+
+  ReadableByteStreamControllerCallPullInto(controller);
+  ReadableByteStreamControllerCallPullOrPullIntoRepeatedlyIfNeeded(controller);
+
+  return promise;
+}
+
+function ReadableByteStreamControllerCallPull(controller) {
+  controller._pullAgain = false;
+  controller._pulling = true;
+
+  try {
+    InvokeOrNoop(controller._underlyingByteSource, 'pull', []);
+  } catch (e) {
+    if (controller._controlledReadableByteStream._state === 'readable') {
+      ErrorReadableByteStreamController(controller, e);
+    }
+  }
+
+  controller._pulling = false;
+}
+
+function ReadableByteStreamControllerCallPullInto(controller) {
+  assert(controller._pendingPullIntos.length > 0);
+  const pullIntoDescriptor = controller._pendingPullIntos[0];
+
+  controller._pullAgain = false;
+  controller._pulling = true;
+
+  try {
+    InvokeOrNoop(controller._underlyingByteSource,
+                 'pullInto',
+                 [new Uint8Array(pullIntoDescriptor.buffer,
+                                 pullIntoDescriptor.byteOffset + pullIntoDescriptor.bytesFilled,
+                                 pullIntoDescriptor.byteLength - pullIntoDescriptor.bytesFilled)]);
+  } catch (e) {
+    if (controller._controlledReadableByteStream._state === 'readable') {
+      ErrorReadableByteStreamController(controller, e);
+    }
+  }
+
+  controller._pulling = false;
+}
+
+function ReadableByteStreamControllerCallPullOrPullIntoLaterIfNeeded(controller) {
+  controller._pullAgain = true;
+
+  if (controller._pulling) {
+    return;
+  }
+
+  process.nextTick(ReadableByteStreamControllerCallPullOrPullIntoRepeatedlyIfNeeded.bind(undefined, controller));
+}
+
+function ReadableByteStreamControllerCallPullOrPullIntoRepeatedlyIfNeeded(controller) {
+  const stream = controller._controlledReadableByteStream;
+
+  while (true) {
+    if (!controller._pullAgain) {
+      return;
+    }
+
+    if (controller._closeRequested) {
+      return;
+    }
+    if (stream._state !== 'readable') {
+      return;
+    }
+
+    if (ReadableStreamHasReader(stream)) {
+      if (GetNumReadRequests(stream) === 0) {
+        return;
+      }
+
+      ReadableByteStreamControllerCallPull(controller);
+    } else if (ReadableStreamHasByobReader(stream)) {
+      if (GetNumReadIntoRequests(stream) === 0) {
+        return;
+      }
+
+      ReadableByteStreamControllerCallPullInto(controller);
+    }
+  }
+}
+
+function RespondToByobReaderInClosedState(controller, buffer) {
+  const firstDescriptor = controller._pendingPullIntos[0];
+
+  if (buffer !== undefined) {
+    firstDescriptor.buffer = buffer;
+  }
+
+  firstDescriptor.buffer = TransferArrayBuffer(firstDescriptor.buffer);
+
+  assert(firstDescriptor.bytesFilled === 0, 'bytesFilled must be 0');
+
+  const stream = controller._controlledReadableByteStream;
+
+  while (GetNumReadIntoRequests(stream) > 0) {
+    const descriptor = controller._pendingPullIntos.shift();
+    RespondToReadIntoRequest(stream, descriptor.buffer);
+  }
+
+  ReleaseReadableStreamReaderGenericForController(controller._controlledReadableByteStream);
+}
+
+function RespondToByobReaderInReadableState(controller, bytesWritten, buffer) {
+  const pullIntoDescriptor = controller._pendingPullIntos[0];
+
+  if (pullIntoDescriptor.bytesFilled + bytesWritten > pullIntoDescriptor.byteLength) {
+    throw new RangeError('bytesWritten out of range');
+  }
+
+  if (buffer !== undefined) {
+    pullIntoDescriptor.buffer = buffer;
+  }
+
+  pullIntoDescriptor.bytesFilled += bytesWritten;
+
+  if (pullIntoDescriptor.bytesFilled < pullIntoDescriptor.elementSize) {
+    // TODO: Figure out whether we should detach the buffer or not here.
+    ReadableByteStreamControllerCallPullOrPullIntoLaterIfNeeded(controller);
+    return;
+  }
+
+  controller._pendingPullIntos.shift();
+
+  const remainderSize = pullIntoDescriptor.bytesFilled % pullIntoDescriptor.elementSize;
+  if (remainderSize > 0) {
+    const end = pullIntoDescriptor.byteOffset + pullIntoDescriptor.bytesFilled;
+    const remainder = pullIntoDescriptor.buffer.slice(end - remainderSize, end);
+    EnqueueChunkToQueueOfController(controller, remainder, 0, remainder.byteLength);
+  }
+
+  RespondToReadIntoRequest(
+      controller._controlledReadableByteStream,
+      TransferArrayBuffer(pullIntoDescriptor.buffer),
+      pullIntoDescriptor.bytesFilled - remainderSize);
+
+  RespondToReadIntoRequestsFromQueue(controller);
+}
+
+function RespondToReadIntoRequestsFromQueue(controller) {
+  assert(!controller._closeRequested);
+
+  while (controller._pendingPullIntos.length > 0) {
+    if (controller._totalQueuedBytes === 0) {
+      ReadableByteStreamControllerCallPullOrPullIntoLaterIfNeeded(controller);
+      return;
+    }
+
+    const pullIntoDescriptor = controller._pendingPullIntos[0];
+
+    const ready = FillPullIntoDescriptorFromQueue(controller, pullIntoDescriptor);
+
+    if (ready) {
+      controller._pendingPullIntos.shift();
+
+      RespondToReadIntoRequest(controller._controlledReadableByteStream, pullIntoDescriptor.buffer, pullIntoDescriptor.bytesFilled);
+    }
+  }
+}
+
+// Other abstract operations
+
+function TransferArrayBuffer(buffer) {
+  // No-op. Just for marking places where detaching an ArrayBuffer is required.
+
+  return buffer;
 }
