+++ reference-implementation/run-tests.js
@@ -6,14 +6,12 @@ import WritableStream from './lib/writable-stream';
 import ByteLengthQueuingStrategy from './lib/byte-length-queuing-strategy';
 import CountQueuingStrategy from './lib/count-queuing-strategy';
 import TransformStream from './lib/transform-stream';
-import ReadableByteStream from './lib/readable-byte-stream';
 
 global.ReadableStream = ReadableStream;
 global.WritableStream = WritableStream;
 global.ByteLengthQueuingStrategy = ByteLengthQueuingStrategy;
 global.CountQueuingStrategy = CountQueuingStrategy;
 global.TransformStream = TransformStream;
-global.ReadableByteStream = ReadableByteStream;
 
 
 if (process.argv.length === 2) {
